+++ src/command_modules/azure-cli-consumption/azure/cli/command_modules/consumption/custom.py
@@ -30,10 +30,10 @@ def cli_consumption_list_usage(client, billing_period_name=None, top=None, inclu
         return list(client.list_by_billing_period(expand=expand, filter=filter_expression))
     elif not billing_period_name and top:
         return list(client.list(expand=expand, filter=filter_expression, top=top).advance_page())
-    return list(client.list(expand=expand, filter=filter_expression, top=top).advance_page())
+    return client.list(expand=expand, filter=filter_expression)
 
 
-def cli_consumption_list_reservations_summaries(client, grain, reservation_order_id, reservation_id=None, start_date=None, end_date=None):
+def cli_consumption_list_reservation_summary(client, grain, reservation_order_id, reservation_id=None, start_date=None, end_date=None):
     filter_from = None
     filter_to = None
     filter_expression = None
@@ -50,7 +50,7 @@ def cli_consumption_list_reservations_summaries(client, grain, reservation_order
     return list(client.list_by_reservation_order(reservation_order_id, grain=grain))
 
 
-def cli_consumption_list_reservations_details(client, reservation_order_id, start_date, end_date, reservation_id=None):
+def cli_consumption_list_reservation_detail(client, reservation_order_id, start_date, end_date, reservation_id=None):
     filter_from = None
     filter_to = None
     filter_expression = None
@@ -71,3 +71,21 @@ def cli_consumption_list_pricesheet_show(client, billing_period_name=None, inclu
     if billing_period_name:
         return client.get_by_billing_period(billing_period_name, expand=expand_properties)
     return client.get(expand=expand_properties)
+
+
+def cli_consumption_list_marketplace(client, billing_period_name=None, start_date=None, end_date=None, top=None):
+    filter_from = None
+    filter_to = None
+    filter_expression = None
+    if start_date and end_date:
+        filter_from = "properties/usageEnd ge \'{}\'".format(start_date.strftime("%Y-%m-%dT%H:%M:%SZ"))
+        filter_to = "properties/usageEnd le \'{}\'".format(end_date.strftime("%Y-%m-%dT%H:%M:%SZ"))
+        filter_expression = "{} and {}".format(filter_from, filter_to)
+
+    if billing_period_name and top:
+        return list(client.list_by_billing_period(billing_period_name, filter=filter_expression, top=top).advance_page())
+    elif billing_period_name and not top:
+        return list(client.list_by_billing_period(billing_period_name, filter=filter_expression))
+    elif not billing_period_name and top:
+        return list(client.list(filter=filter_expression, top=top).advance_page())
+    return client.list(filter=filter_expression)