+++ tools/automation/cli_linter/rules/parameter_rules.py
@@ -3,4 +3,11 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-"""Command linting automation code"""
+from ..rule_decorators import parameter_rule
+from ..linter import RuleError
+
+
+@parameter_rule
+def missing_parameter_help_rule(linter, command_name, parameter_name):
+    if not linter.get_parameter_help(command_name, parameter_name):
+        raise RuleError('Missing help')