+++ src/command_modules/azure-cli-role/HISTORY.rst
@@ -2,6 +2,16 @@
 
 Release History
 ===============
+
+2.0.23
+++++++
+* BREAKING CHANGE: remove deprecated `az ad sp reset-credentials`
+* Minor fixes.
+
+2.0.22
+++++++
+* `sdist` is now compatible with wheel 0.31.0
+
 2.0.21
 ++++++
 * graph: support required access configuration and native client